+++ __tests__/cairo1.test.ts
@@ -1,8 +1,12 @@
 import {
   Account,
+  BigNumberish,
   CallData,
+  Calldata,
   Contract,
   DeclareDeployUDCResponse,
+  RawArgsArray,
+  RawArgsObject,
   SequencerProvider,
   cairo,
   ec,
@@ -14,6 +18,7 @@ import {
 import {
   compiledC1Account,
   compiledC1AccountCasm,
+  compiledComplexSierra,
   compiledHelloSierra,
   compiledHelloSierraCasm,
   describeIfDevnetSequencer,
@@ -107,8 +112,13 @@ describeIfDevnetSequencer('Cairo 1 Devnet Sequencer', () => {
     });
 
     test('Cairo 1 - uint256', async () => {
+      // defined as number
       const result = await cairo1Contract.test_u256(2n ** 256n - 2n);
       expect(result).toBe(2n ** 256n - 1n);
+
+      // defined as struct
+      const result1 = await cairo1Contract.test_u256(cairo.uint256(2n ** 256n - 2n));
+      expect(result1).toBe(2n ** 256n - 1n);
     });
 
     test('Cairo 1 Contract Interaction - bool', async () => {
@@ -151,8 +161,18 @@ describeIfDevnetSequencer('Cairo 1 Devnet Sequencer', () => {
       const status = await cairo1Contract.echo_array([123, 55, 77, 255]);
       expect(status).toEqual([123n, 55n, 77n, 255n]);
 
+      // uint256 defiend as number
       const status1 = await cairo1Contract.echo_array_u256([123, 55, 77, 255]);
       expect(status1).toEqual([123n, 55n, 77n, 255n]);
+
+      // uint256 defined as struct
+      const status11 = await cairo1Contract.echo_array_u256([
+        cairo.uint256(123),
+        cairo.uint256(55),
+        cairo.uint256(77),
+        cairo.uint256(255),
+      ]);
+      expect(status11).toEqual([123n, 55n, 77n, 255n]);
 
       const status2 = await cairo1Contract.echo_array_bool([true, true, false, false]);
       expect(status2).toEqual([true, true, false, false]);
@@ -237,6 +257,151 @@ describeIfDevnetSequencer('Cairo 1 Devnet Sequencer', () => {
       const res = await cairo1Contract.mix_req([1, 2, 3], true);
       expect(res).toEqual([1n, 2n, 3n, 1n, 2n]);
     });
+
+    test('myCallData.compile for Cairo 1', async () => {
+      const myFalseUint256 = { high: 1, low: 23456 }; // wrong order
+      type Order2 = {
+        p1: BigNumberish;
+        p2: BigNumberish[];
+      };
+
+      const myOrder2bis: Order2 = {
+        // wrong order
+        p2: [234, 467456745457n, '0x56ec'],
+        p1: '17',
+      };
+      const myRawArgsObject: RawArgsObject = {
+        // wrong order
+        active: true,
+        symbol: 'NIT',
+        initial_supply: myFalseUint256,
+        recipient: '0x7e00d496e324876bbc8531f2d9a82bf154d1a04a50218ee74cdd372f75a551a',
+        decimals: 18,
+        tupoftup: cairo.tuple(cairo.tuple(34, '0x5e'), myFalseUint256),
+        card: myOrder2bis,
+        longText: 'Bug is back, for ever, here and everywhere',
+        array1: [100, 101, 102],
+        array2: [
+          [200, 201],
+          [202, 203],
+          [204, 205],
+        ],
+        array3: [myOrder2bis, myOrder2bis],
+        array4: [myFalseUint256, myFalseUint256],
+        tuple1: cairo.tuple(40000n, myOrder2bis, [54, 55n, '0xae'], 'texte'),
+        name: 'niceToken',
+        array5: [cairo.tuple(251, 40000n), cairo.tuple(252, 40001n)],
+      };
+      const myRawArgsArray: RawArgsArray = [
+        'niceToken',
+        'NIT',
+        18,
+        { low: 23456, high: 1 },
+        { p1: '17', p2: [234, 467456745457n, '0x56ec'] },
+        '0x7e00d496e324876bbc8531f2d9a82bf154d1a04a50218ee74cdd372f75a551a',
+        true,
+        { '0': { '0': 34, '1': '0x5e' }, '1': { low: 23456, high: 1 } },
+        'Bug is back, for ever, here and everywhere',
+        [100, 101, 102],
+        [
+          [200, 201],
+          [202, 203],
+          [204, 205],
+        ],
+        [
+          { p1: '17', p2: [234, 467456745457n, '0x56ec'] },
+          { p1: '17', p2: [234, 467456745457n, '0x56ec'] },
+        ],
+        [
+          { low: 23456, high: 1 },
+          { low: 23456, high: 1 },
+        ],
+        {
+          '0': 40000n,
+          '1': { p1: '17', p2: [234, 467456745457n, '0x56ec'] },
+          '2': [54, 55n, '0xae'],
+          '3': 'texte',
+        },
+        [
+          { '0': 251, '1': 40000n },
+          { '0': 252, '1': 40001n },
+        ],
+      ];
+
+      const contractCallData: CallData = new CallData(compiledComplexSierra.abi);
+      const callDataFromObject: Calldata = contractCallData.compile('constructor', myRawArgsObject);
+      const callDataFromArray: Calldata = contractCallData.compile('constructor', myRawArgsArray);
+      const expectedResult = [
+        '2036735872918048433518',
+        '5130580',
+        '18',
+        '23456',
+        '1',
+        '17',
+        '3',
+        '234',
+        '467456745457',
+        '22252',
+        '3562055384976875123115280411327378123839557441680670463096306030682092229914',
+        '1',
+        '34',
+        '94',
+        '23456',
+        '1',
+        '2',
+        '117422190885827407409664260607192623408641871979684112605616397634538401380',
+        '39164769268277364419555941',
+        '3',
+        '100',
+        '101',
+        '102',
+        '3',
+        '2',
+        '200',
+        '201',
+        '2',
+        '202',
+        '203',
+        '2',
+        '204',
+        '205',
+        '2',
+        '17',
+        '3',
+        '234',
+        '467456745457',
+        '22252',
+        '17',
+        '3',
+        '234',
+        '467456745457',
+        '22252',
+        '2',
+        '23456',
+        '1',
+        '23456',
+        '1',
+        '40000',
+        '0',
+        '17',
+        '3',
+        '234',
+        '467456745457',
+        '22252',
+        '3',
+        '54',
+        '55',
+        '174',
+        '499918599269',
+        '2',
+        '251',
+        '40000',
+        '252',
+        '40001',
+      ];
+      expect(callDataFromObject).toStrictEqual(expectedResult);
+      expect(callDataFromArray).toStrictEqual(expectedResult);
+    });
   });
 
   describe('Cairo1 Account contract', () => {
@@ -252,11 +417,13 @@ describeIfDevnetSequencer('Cairo 1 Devnet Sequencer', () => {
       const calldata = { publicKey: pubKey };
 
       // declare account
-      const declareAccount = await account.declare({
+      const declareAccount = await account.declareIfNot({
         contract: compiledC1Account,
         casm: compiledC1AccountCasm,
       });
-      await account.waitForTransaction(declareAccount.transaction_hash);
+      if (declareAccount.transaction_hash) {
+        await account.waitForTransaction(declareAccount.transaction_hash);
+      }
       const accountClassHash = declareAccount.class_hash;
 
       // fund new account
@@ -279,7 +446,7 @@ describeIfDevnetSequencer('Cairo 1 Devnet Sequencer', () => {
       await account.waitForTransaction(transaction_hash);
 
       // deploy account
-      accountC1 = new Account(provider, toBeAccountAddress, priKey);
+      accountC1 = new Account(provider, toBeAccountAddress, priKey, '1');
       const deployed = await accountC1.deploySelf({
         classHash: accountClassHash,
         constructorCalldata: calldata,