+++ src/utils/transaction.ts
@@ -1,7 +1,7 @@
-import { CairoVersion, Call, CallStruct, Calldata, ParsedStruct } from '../types';
+import { BigNumberish, CairoVersion, Call, CallStruct, Calldata, ParsedStruct } from '../types';
 import { CallData } from './calldata';
 import { getSelectorFromName } from './hash';
-import { BigNumberish, toBigInt } from './num';
+import { toBigInt } from './num';
 
 /**
  * Transforms a list of Calls, each with their own calldata, into
@@ -59,7 +59,6 @@ export const transformCallsToMulticallArrays_cairo1 = (calls: Call[]) => {
   return callArray;
 };
 
-// TT: Can be removed ?
 /**
  * Transforms a list of calls in the full flattened calldata expected
  * by the __execute__ protocol.
@@ -67,7 +66,14 @@ export const transformCallsToMulticallArrays_cairo1 = (calls: Call[]) => {
  * @returns Calldata
  */
 export const fromCallsToExecuteCalldata_cairo1 = (calls: Call[]) => {
-  return CallData.compile({ calls });
+  // ensure property order
+  const orderCalls = calls.map((call) => ({
+    contractAddress: call.contractAddress,
+    entrypoint: call.entrypoint,
+    calldata: call.calldata,
+  }));
+
+  return CallData.compile({ orderCalls });
 };
 
 /**
@@ -78,7 +84,7 @@ export const fromCallsToExecuteCalldata_cairo1 = (calls: Call[]) => {
  */
 export const getExecuteCalldata = (calls: Call[], cairoVersion: CairoVersion = '0') => {
   if (cairoVersion === '1') {
-    return CallData.compile({ calls });
+    return fromCallsToExecuteCalldata_cairo1(calls);
   }
   return fromCallsToExecuteCalldata(calls);
 };